feat: create named export for createEvent
